fix format and add root scripts
