+++ .prettierignore
@@ -4,3 +4,4 @@ dist
 node_modules
 package-lock.json
 css
+*.min.js
