+++ packages/graphiql/.prettierignore
@@ -1,0 +1,4 @@
+*.min.*
+bundle/*
+dist
+
