+++ packages/graphiql/package.json
@@ -26,6 +26,7 @@
     "bundle": "pnpm bundle:webpack && pnpm bundle:ts",
     "bundle:ts": "node scripts/build-graphiql.cjs",
     "bundle:webpack": "webpack -c browser/webpack.config.js && cpy browser/css/graphiql.min.css bundle/",
+    "lint": "prettier --check .",
     "prepack": "bob-esbuild build",
     "prepublishOnly": "pnpm bundle && bob-esbuild build",
     "postpublish": "gh-release"
